+++ Extension/src/pages/fullscreen-user-rules/components/fullscreen-user-rules/FullscreenUserRules.tsx
@@ -40,13 +40,15 @@ import '../../../options/components/UserRules/styles.pcss';
 export const FullscreenUserRules = observer(() => {
     const store = useContext(fullscreenUserRulesStore);
 
-    const { uiStore: { showLoader } } = useContext(rootStore);
+    // TODO: Move rootStore fields, that fullscreenUserRules need to fullscreenUserRulesStore: AG-48937
+    const { uiStore: { showLoader }, settingsStore } = useContext(rootStore);
 
     useAppearanceTheme(store.appearanceTheme);
 
     // append message listeners
     useEffect(() => {
         store.getFullscreenUserRulesData();
+        settingsStore.getFullscreenUserRulesData();
 
         let removeListenerCallback = () => {};
 
@@ -78,7 +80,7 @@ export const FullscreenUserRules = observer(() => {
         return () => {
             removeListenerCallback();
         };
-    }, [store]);
+    }, [store, settingsStore]);
 
     return (
         <>