+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Apply $document blocking rule even if specific exception rule is present [#3262].
 - Blocked iframes are not collapsed on Firefox [#3116].
 - The extension got an unexpected error on [::]:8000 [#3360]
+- Toggle, that doesn't work in User Rules fullscreen mode [#3365]
 
 [#2934]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2934
 [#3069]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3069
@@ -42,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3278]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3278
 [#3262]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3262
 [#3360]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3360
+[#3365]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3365
 
 ## [v5.2.300.0] - ??